pyproject.toml
[project]
name = "s3-data-lake-mcp-server"
version = "1.0.0"
description = "MCP server for reading and querying data from S3 bucket data lakes"
authors = [
    {name = "AWS AgentCore Team", email = "agentcore@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mcp", "s3", "data-lake", "aws", "agentcore", "bedrock"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mcp>=1.0.0",
    "boto3>=1.34.0",
    "pandas>=2.0.0",
    "pyarrow>=10.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "starlette>=0.27.0",
    "botocore>=1.42.19",
    "requests-aws4auth>=1.3.1",
    "faker>=40.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
test = [
    "strands-agents[mcp]>=0.1.0",
    "moto[s3]>=4.0.0",
]

[project.scripts]
s3-mcp-server = "s3_data_lake_mcp_server:main"
upload-sample-data = "upload_sample_data:main"

[project.urls]
Homepage = "https://github.com/aws/s3-data-lake-mcp-server"
Documentation = "https://github.com/aws/s3-data-lake-mcp-server#readme"
Repository = "https://github.com/aws/s3-data-lake-mcp-server"
Issues = "https://github.com/aws/s3-data-lake-mcp-server/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
